serving/requirements.txt
flask
torch
transformers
underthesea
python-dotenv
pyngrok
